problems with optional support :(
