+++ CMakeLists.txt
@@ -74,6 +74,8 @@ if(MPACK_CPP_BUILD_TESTS)
         tests/test_mpack_cpp.cpp
         tests/test_nested_struct.cpp
         tests/test_custom_allocator.cpp
+        tests/test_optional.cpp
+        tests/test_ext_type.cpp
     )
     target_link_libraries(
         test_mpack_cpp
